Delete some warnings

+++ lib/GAME_H/obstacle.hpp
@@ -16,8 +16,8 @@ private:
 public:
     Obstacle(glm::vec3 pos);
     void draw(glm::mat4 view, glm::mat4 projection, glm::mat4 model, glm::vec3 camPos, SceneLight sceneLight, std::vector<Light*> lights) const;
-    int getPosX() const {return m_pos.x;};
-    int getPosZ() const {return m_pos.z;};
+    float getPosX() const {return m_pos.x;};
+    float getPosZ() const {return m_pos.z;};
     //fonction de collision  
 };
